move struct to utils directory.
