+++ connection/connection.go
@@ -1,1 +1,1 @@
-package pbspro
+package connection
