+++ qstat/qstat.go
@@ -1,1 +1,1 @@
-package pbspro
+package qstat
